create folder move files and make some methods protected (#177)
* create folder move files and make some methods protected
