+++ src/test/java/io/confluent/idesidecar/restapi/cache/SchemaRegistryClientsTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
+import io.confluent.idesidecar.restapi.clients.SchemaRegistryClients;
 import io.confluent.kafka.schemaregistry.client.MockSchemaRegistryClient;
 import io.confluent.kafka.schemaregistry.client.SchemaRegistryClient;
 import io.quarkus.test.junit.QuarkusTest;
